add diagrams, refactor verifiableCredential service into hexagonal, update credential offer
