+++ docs/diagrams/credential-emission.puml
@@ -3,18 +3,24 @@
 autonumber
 
 boundary "Wallet" as wallet
-boundary "Issuer VCI" as issuerVCI
-boundary "Authentic Sources" as authSources
-boundary "Remote Signature" as remoteSignature
+boundary "Credential Issuer" as issuerVCI
+boundary "DSS/Remote Signature" as remoteSignature
+boundary "Auth Server" as authserver
 
-wallet -> wallet++: Create Credential Request
-wallet -> wallet: Build dynamic ur to get Credential
-wallet -> issuerVCI++: Request Credential \n POST /credentials
+issuerVCI -> issuerVCI: Generate Credential Offer
+issuerVCI <-> authserver: Get Pre Authorization Code
+wallet <- issuerVCI: Scan QR with Credential Offer Uri
+wallet -> issuerVCI: Execute QR content \n GET /credential-offer/{nonce}
+issuerVCI -> wallet: Credential Offer
+wallet -> authserver: Get token from pre-auth-code
+authserver <-> issuerVCI: Save Token and get nonce
+authserver -> wallet: Return Token and nonce
+wallet -> issuerVCI: Get Credential
 issuerVCI -> issuerVCI: Get the KID of header proof \n (did:key)
 issuerVCI -> issuerVCI: Get the format \n (jwt_vc_json)
-issuerVCI -> authSources++: Request user data
-authSources -> issuerVCI--: 200 OK user_data
+issuerVCI -> issuerVCI: Get user data from local file
 issuerVCI -> issuerVCI: Generate the Credential \n (from credential_template + user_data)
+issuerVCI <-> remoteSignature: Get DID from DSS/Remote Signature
 issuerVCI -> remoteSignature++: Sign Credential
 remoteSignature -> remoteSignature: Get signature from Digital Certificate
 remoteSignature -> remoteSignature: Sign the Credential with JAdES
