+++ docs/diagrams/credential-offer.puml
@@ -4,8 +4,8 @@ autonumber
 
 boundary "Wallet" as wallet
 boundary "Issuer SPA" as issuerSpa
-boundary "Issuer VCI" as issuerVci
-boundary "Auth OIDC" as authOidc
+boundary "Credential Issuer" as issuerVci
+boundary "Auth Server" as authOidc
 
 issuerSpa -> issuerVci++: GET https://<domain>/credential-offers \n"Authorization":"access_token"
 issuerVci -> issuerVci: Creation Credential Offer
@@ -20,9 +20,9 @@ issuerVci -> issuerSpa--: 200 OK Response \n{"data":"credential_offer_uri"}
 issuerSpa -> issuerSpa: Show QR code
 wallet <-> issuerSpa: Scan QR code
 wallet -> wallet: Execute "credential_offer_uri"
-wallet -> issuerVci++: GET https://<domain>/credential-offers/{id}
+wallet -> issuerVci++: GET https://<domain>/credential-offer/{id}
 issuerVci -> issuerVci: Get credential offer by ID
-issuerVci -> wallet: 200 OK \n{"data":"credential_offer"}
+issuerVci -> wallet--: 200 OK \n{"data":"credential_offer"}
 
 
 @enduml
