+++ src/main/java/es/in2/issuer/application/service/VerifiableCredentialIssuanceService.java
@@ -1,0 +1,9 @@
+package es.in2.issuer.application.service;
+
+import es.in2.issuer.domain.model.CredentialRequest;
+import es.in2.issuer.domain.model.VerifiableCredentialResponse;
+import reactor.core.publisher.Mono;
+
+public interface VerifiableCredentialIssuanceService {
+    Mono<VerifiableCredentialResponse> generateVerifiableCredentialResponse(String username, CredentialRequest credentialRequest, String token);
+}
